refactor: split stats planner helpers
